reorganising and splitting complex parts
