+++ .gitignore
@@ -1,4 +1,5 @@
 .idea
 node_modules
+src/**/*.js
 src/*.js
 /tsconfig.tsbuildinfo
