+++ src/configuration/file-builder.ts
@@ -1,0 +1,8 @@
+export interface FileBuilder {
+  buildIndex: boolean,
+  minifyPages: boolean,
+  appendPageChunk: boolean,
+  preLoadCSS: boolean,
+  preLoadJS: boolean,
+  preloadStartDelay: number,
+}
