+++ src/configuration/file-finder.ts
@@ -1,0 +1,5 @@
+export interface FileFinder {
+  fileName: string,
+  pagesRoot: string,
+  distJSRoot: string,
+}
