+++ src/configuration/html-minify.ts
@@ -1,0 +1,7 @@
+export interface HTMLMinify {
+  collapseBooleanAttributes: boolean,
+  conservativeCollapse: boolean,
+  collapseWhitespace: boolean,
+  removeAttributeQuotes: boolean,
+  removeComments: boolean,
+}
