+++ src/configuration/routes.ts
@@ -1,0 +1,8 @@
+export interface Routes {
+  build: boolean,
+  type: 'tsx' | 'jsx',
+  overridePathMappings: {
+    [filesystemPath: string]: string,
+  },
+  useOutput: boolean,
+}
