+++ src/configuration/runtime.ts
@@ -1,0 +1,3 @@
+export interface Runtime {
+  reloadWaitMs: number,
+}
