+++ src/configuration/sitemap.ts
@@ -1,0 +1,4 @@
+export interface Sitemap {
+  domain: string,
+  build: boolean,
+}
